feature: use go-getter for all downloads with available progress tracking
